# IO.Swagger/Controllers/ConsultarApi.cs
/*
 * Web API Tareas
 *
 * Servicio de Gestión de Tareas: proporciona los métodos necesarios para la consulta, creación, actualización y borrado de las tareas a los usuarios autorizados.
 *
 * OpenAPI spec version: 1.0
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Collections.Generic;
using System.Linq;
using System.Net;
using System.Threading.Tasks;
using Microsoft.AspNetCore.Http;
using Microsoft.AspNetCore.Mvc;
using Microsoft.AspNetCore.WebUtilities;
using Microsoft.Extensions.Logging;
using Microsoft.Extensions.Primitives;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using IO.Swagger.Attributes;
using IO.Swagger.Models;

namespace IO.Swagger.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    public class ConsultarApiController : Controller
    { 
        /// <summary>
        /// Consultar Tareas
        /// </summary>
        /// <remarks>Lista todas las Tareas ordenadas por Fecha de Vencimiento. Permite filtrar por Usuario y/o por el Indicador de si la tarea está finalizada</remarks>
        /// <param name="usuario">Usuario dueño de la Tarea</param>
        /// <param name="finalizada">Indicador de si la tarea se encuentra finalizada</param>
        /// <response code="200">successful operation</response>
        [HttpGet]
        [Route("/api/Tareas")]
        [ValidateModelState]
        [SwaggerOperation("GetTareas")]
        [SwaggerResponse(statusCode: 200, type: typeof(List<Tarea>), description: "successful operation")]
        public virtual IActionResult GetTareas([FromQuery]string usuario, [FromQuery]string finalizada)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(List<Tarea>));
            string exampleJson = null;
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<List<Tarea>>(exampleJson)
                        : default(List<Tarea>);            //TODO: Change the data returned
            return new ObjectResult(example);
        }
    }
}