# IO.Swagger/Controllers/CrearApi.cs
/*
 * Web API Tareas
 *
 * Servicio de Gestión de Tareas: proporciona los métodos necesarios para la consulta, creación, actualización y borrado de las tareas a los usuarios autorizados.
 *
 * OpenAPI spec version: 1.0
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Collections.Generic;
using System.Linq;
using System.Net;
using System.Threading.Tasks;
using Microsoft.AspNetCore.Http;
using Microsoft.AspNetCore.Mvc;
using Microsoft.AspNetCore.WebUtilities;
using Microsoft.Extensions.Logging;
using Microsoft.Extensions.Primitives;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using IO.Swagger.Attributes;
using IO.Swagger.Models;

namespace IO.Swagger.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    public class CrearApiController : Controller
    { 
        /// <summary>
        /// Crear Tarea
        /// </summary>
        /// <remarks>Realiza la creación de una Tarea</remarks>
        /// <param name="body">Tarea. Para la creación de Tarea no se requiere el atributo Id</param>
        /// <response code="200">successful operation</response>
        /// <response code="400">Bad Request Response (Los parametros descripcion y fecha_vencimiento de la Tarea son obligatorios)</response>
        [HttpPost]
        [Route("/api/Tareas/crear")]
        [ValidateModelState]
        [SwaggerOperation("PostTarea")]
        [SwaggerResponse(statusCode: 200, type: typeof(Tarea), description: "successful operation")]
        [SwaggerResponse(statusCode: 400, type: typeof(string), description: "Bad Request Response (Los parametros descripcion y fecha_vencimiento de la Tarea son obligatorios)")]
        public virtual IActionResult PostTarea([FromBody]TareaCrear body)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(Tarea));

            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400, default(string));
            string exampleJson = null;
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<Tarea>(exampleJson)
                        : default(Tarea);            //TODO: Change the data returned
            return new ObjectResult(example);
        }
    }
}