# IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.2",
  "info" : {
    "title" : "Web API Tareas",
    "description" : "Servicio de Gestión de Tareas: proporciona los métodos necesarios para la consulta, creación, actualización y borrado de las tareas a los usuarios autorizados.",
    "version" : "1.0"
  },
  "servers" : [ {
    "url" : "http://localhost:56000/api"
  } ],
  "security" : [ {
    "JWT_token" : [ ]
  } ],
  "paths" : {
    "/login/autorizar" : {
      "post" : {
        "tags" : [ "Login" ],
        "summary" : "Generar JWT Token de conexión",
        "description" : "Realiza la generación del JWT Token a partir de los datos de usuario",
        "operationId" : "Authenticate",
        "requestBody" : {
          "description" : "LoginRequest",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "title" : "Token",
                  "type" : "string",
                  "example" : "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1bmlxdWVfbmFtZSI6InVzZXIxIiwibmJmIjoxNTUzMjI1MTMzLCJleHAiOjE1NTMyMjY5MzMsImlhdCI6MTU1MzIyNTEzMywiaXNzIjoiaHR0cDovL2xvY2FsaG9zdDo1NjAwMCIsImF1ZCI6Imh0dHA6Ly9sb2NhbGhvc3Q6NTYwMDAifQ.jRep6JVq5urnIOsiPBgCS9P-ljPtuhqQe5vzx7OKRos"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request Response (Los datos de login son obligatorios)",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "title" : "Los datos de login son obligatorios",
                  "type" : "string",
                  "example" : "Bad Request"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized (Los datos de login son erróneos)",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "title" : "Los datos de login son erróneos",
                  "type" : "string",
                  "example" : "Unauthorized"
                }
              }
            }
          }
        }
      }
    },
    "/Tareas" : {
      "get" : {
        "tags" : [ "Consultar" ],
        "summary" : "Consultar Tareas",
        "description" : "Lista todas las Tareas ordenadas por Fecha de Vencimiento. Permite filtrar por Usuario y/o por el Indicador de si la tarea está finalizada",
        "operationId" : "GetTareas",
        "parameters" : [ {
          "name" : "usuario",
          "in" : "query",
          "description" : "Usuario dueño de la Tarea",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          },
          "example" : "usuario12345"
        }, {
          "name" : "finalizada",
          "in" : "query",
          "description" : "Indicador de si la tarea se encuentra finalizada",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string",
            "enum" : [ "true", "false" ]
          },
          "example" : "true"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Tarea"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/Tareas/crear" : {
      "post" : {
        "tags" : [ "Crear" ],
        "summary" : "Crear Tarea",
        "description" : "Realiza la creación de una Tarea",
        "operationId" : "PostTarea",
        "requestBody" : {
          "description" : "Tarea. Para la creación de Tarea no se requiere el atributo Id",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Tarea_Crear"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Tarea"
                }
              }
            }
          },
          "400" : {
            "description" : "Bad Request Response (Los parametros descripcion y fecha_vencimiento de la Tarea son obligatorios)",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "title" : "Los parametros descripcion y fecha_vencimiento de la Tarea son obligatorios",
                  "type" : "string",
                  "example" : "Bad Request"
                }
              }
            }
          }
        }
      }
    },
    "/Tareas/actualizar" : {
      "post" : {
        "tags" : [ "Actualizar" ],
        "summary" : "Actualizar Tarea",
        "description" : "Realiza la actualización de una Tarea",
        "operationId" : "UpdateTarea",
        "requestBody" : {
          "description" : "Tarea. Para la actualización de Tarea se requiere el atributo Id",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Tarea"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Tarea"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found Response (La Tarea no existe)",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "title" : "La Tarea no existe",
                  "type" : "string",
                  "example" : "Not Found"
                }
              }
            }
          }
        }
      }
    },
    "/Tareas/borrar" : {
      "post" : {
        "tags" : [ "Borrar" ],
        "summary" : "Borrar Tarea",
        "description" : "Realiza el borrado de una Tarea",
        "operationId" : "DeleteTarea",
        "requestBody" : {
          "description" : "Tarea. Para el borrado de Tarea se requiere el atributo Id",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Tarea_Borrar"
              }
            }
          }
        },
        "responses" : {
          "204" : {
            "description" : "successful operation",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "title" : "No Content",
                  "type" : "string",
                  "example" : "No Content"
                }
              }
            }
          },
          "404" : {
            "description" : "Not Found Response (La Tarea no existe)",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "title" : "La Tarea no existe",
                  "type" : "string",
                  "example" : "Not Found"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "Tarea" : {
        "title" : "Tarea",
        "type" : "object",
        "properties" : {
          "Id" : {
            "type" : "number",
            "description" : "Id de la Tarea",
            "example" : 59
          },
          "fecha_creacion" : {
            "type" : "string",
            "description" : "Fecha Creación de la Tarea",
            "format" : "date-time"
          },
          "descripcion" : {
            "type" : "string",
            "description" : "Descripción de la Tarea",
            "example" : "Tarea No 1"
          },
          "ind_finalizada" : {
            "type" : "boolean",
            "description" : "Indicador de si la Tarea se encuentra finalizada (true/false)",
            "example" : false
          },
          "fecha_vencimiento" : {
            "type" : "string",
            "description" : "Fecha Vencimiento de la Tarea",
            "format" : "date-time"
          },
          "usuario" : {
            "type" : "string",
            "description" : "Usuario dueño de la Tarea",
            "example" : "user12345"
          }
        }
      },
      "Tarea_Crear" : {
        "title" : "Tarea (Crear)",
        "type" : "object",
        "properties" : {
          "fecha_creacion" : {
            "type" : "string",
            "description" : "Fecha Creación de la Tarea",
            "format" : "date-time"
          },
          "descripcion" : {
            "type" : "string",
            "description" : "Descripción de la Tarea",
            "example" : "Tarea No 1"
          },
          "ind_finalizada" : {
            "type" : "boolean",
            "description" : "Indicador de si la Tarea se encuentra finalizada(true/false)",
            "example" : false
          },
          "fecha_vencimiento" : {
            "type" : "string",
            "description" : "Fecha Vencimiento de la Tarea",
            "format" : "date-time"
          },
          "usuario" : {
            "type" : "string",
            "description" : "Usuario dueño de la Tarea",
            "example" : "user12345"
          }
        }
      },
      "Tarea_Borrar" : {
        "title" : "Tarea (Borrar)",
        "type" : "object",
        "properties" : {
          "Id" : {
            "type" : "number",
            "description" : "Id de la Tarea",
            "example" : 17
          }
        }
      },
      "LoginRequest" : {
        "title" : "LoginRequest",
        "type" : "object",
        "properties" : {
          "Username" : {
            "type" : "string",
            "description" : "Usuario",
            "example" : "usuario12345"
          },
          "Password" : {
            "type" : "string",
            "description" : "Password",
            "example" : "pwd123456"
          }
        }
      }
    },
    "securitySchemes" : {
      "JWT_token" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}